# docs/openapi.json
{"openapi":"3.0.2","info":{"title":"MOVIES","version":"1.0.0"},"paths":{"/api/v1/films/":{"get":{"tags":["Фильмы"],"summary":"Films Popular","description":"Получить популярные фильмы (в текущей версии - с наибольшим рейтингом).\n\n- **sort**: поле для сортировки с префиксом + либо -\n- **filter[genre]**: UUID идентификатор жанра, из которого получить фильмы\n- **page[number]**: номер страницы\n- **page[size]**: количество фильмов на странице","operationId":"films_popular_api_v1_films__get","parameters":[{"required":false,"schema":{"allOf":[{"$ref":"#/components/schemas/Sorting"}],"default":"-imdb_rating"},"name":"sort","in":"query"},{"required":false,"schema":{"title":"Filter[Genre]","type":"string","format":"uuid"},"name":"filter[genre]","in":"query"},{"required":false,"schema":{"title":"Page[Number]","minimum":1.0,"type":"integer","default":1},"name":"page[number]","in":"query"},{"required":false,"schema":{"title":"Page[Size]","minimum":1.0,"type":"integer","default":50,"lte":200},"name":"page[size]","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImdbfilmManyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/films/search/":{"get":{"tags":["Фильмы"],"summary":"Film Search","description":"Найти фильмы.\n\n- **query**: поисковый запрос\n- **filter[genre]**: UUID идентификатор жанра, в котором выполнить поиск\n- **page[number]**: номер страницы\n- **page[size]**: количество фильмов на странице","operationId":"film_search_api_v1_films_search__get","parameters":[{"required":true,"schema":{"title":"Query","type":"string"},"name":"query","in":"query"},{"required":false,"schema":{"title":"Filter[Genre]","type":"string","format":"uuid"},"name":"filter[genre]","in":"query"},{"required":false,"schema":{"title":"Page[Number]","minimum":1.0,"type":"integer","default":1},"name":"page[number]","in":"query"},{"required":false,"schema":{"title":"Page[Size]","minimum":1.0,"type":"integer","default":50,"lte":200},"name":"page[size]","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImdbfilmManyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/films/{film_id}/similar":{"get":{"tags":["Фильмы"],"summary":"Film Similar","description":"Получить похожие фильмы (в текущей версии - фильмы того же жанра).\n\n- **film_id**: UUID идентификатор фильма\n- **page[number]**: номер страницы\n- **page[size]**: количество элементов на странице","operationId":"film_similar_api_v1_films__film_id__similar_get","parameters":[{"required":true,"schema":{"title":"Film Id","type":"string","format":"uuid"},"name":"film_id","in":"path"},{"required":false,"schema":{"title":"Page[Number]","minimum":1.0,"type":"integer","default":1},"name":"page[number]","in":"query"},{"required":false,"schema":{"title":"Page[Size]","minimum":1.0,"type":"integer","default":50,"lte":200},"name":"page[size]","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImdbfilmManyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/films/{film_id}":{"get":{"tags":["Фильмы"],"summary":"Film Details","description":"Получить полную информацию о фильме.\n\n- **film_id**: UUID идентификатор фильма","operationId":"film_details_api_v1_films__film_id__get","parameters":[{"required":true,"schema":{"title":"Film Id","type":"string","format":"uuid"},"name":"film_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExtendedImdbFilm"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/search/{person_id}":{"get":{"tags":["Персоны"],"summary":"Person By Id","operationId":"person_by_id_api_v1_persons_search__person_id__get","parameters":[{"required":true,"schema":{"title":"Person Id","type":"string","format":"uuid"},"name":"person_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExtendedPerson"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/{person_id}/film":{"get":{"tags":["Персоны"],"summary":"Films By Person","operationId":"films_by_person_api_v1_persons__person_id__film_get","parameters":[{"required":true,"schema":{"title":"Person Id","type":"string","format":"uuid"},"name":"person_id","in":"path"},{"required":false,"schema":{"title":"count of results rows","minimum":1.0,"type":"integer","default":50},"name":"page[size]","in":"query"},{"required":false,"schema":{"title":"number of page (pagination)","minimum":1.0,"type":"integer","default":1},"name":"page[number]","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImdbfilmManyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/":{"get":{"tags":["Персоны"],"summary":"Person Search","operationId":"person_search_api_v1_persons__get","parameters":[{"required":false,"schema":{"title":"string for search","type":"string","default":""},"name":"query","in":"query"},{"required":false,"schema":{"title":"count of results rows","minimum":1.0,"type":"integer","default":50},"name":"page[size]","in":"query"},{"required":false,"schema":{"title":"number of page (pagination)","minimum":1.0,"type":"integer","default":1},"name":"page[number]","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExtendedpersonManyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/genres/{genre_id}":{"get":{"tags":["Жанры"],"summary":"Genre By Id","operationId":"genre_by_id_api_v1_genres__genre_id__get","parameters":[{"required":true,"schema":{"title":"Genre Id","type":"string","format":"uuid"},"name":"genre_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Genre"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/genres/":{"get":{"tags":["Жанры"],"summary":"All Genres","operationId":"all_genres_api_v1_genres__get","parameters":[{"required":false,"schema":{"title":"count of results rows","minimum":1.0,"type":"integer","default":50},"name":"page[size]","in":"query"},{"required":false,"schema":{"title":"number of page (pagination)","minimum":1.0,"type":"integer","default":1},"name":"page[number]","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreManyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ExtendedImdbFilm":{"title":"ExtendedImdbFilm","required":["uuid","title","imdb_rating","description","genres","actors","writers","directors"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"title":{"title":"movie title","type":"string"},"imdb_rating":{"title":"IMDB rating","type":"number"},"description":{"title":"movie description","type":"string"},"genres":{"title":"Genres","type":"array","items":{"$ref":"#/components/schemas/Genre"}},"actors":{"title":"Actors","type":"array","items":{"$ref":"#/components/schemas/Person"}},"writers":{"title":"Writers","type":"array","items":{"$ref":"#/components/schemas/Person"}},"directors":{"title":"Directors","type":"array","items":{"$ref":"#/components/schemas/Person"}}},"description":"Film with extended info"},"ExtendedPerson":{"title":"ExtendedPerson","required":["uuid","full_name","movies"],"type":"object","properties":{"uuid":{"title":"person id","type":"string","format":"uuid"},"full_name":{"title":"persons full name","type":"string"},"movies":{"title":"Movies","type":"array","items":{"$ref":"#/components/schemas/RoleMovies"}}},"description":"Person info"},"ExtendedpersonManyResponse":{"title":"ExtendedpersonManyResponse","required":["total","result"],"type":"object","properties":{"total":{"title":"Amount rows in source","type":"integer"},"result":{"title":"Result","type":"array","items":{"$ref":"#/components/schemas/ExtendedPerson"}}}},"Film":{"title":"Film","required":["uuid","title"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"title":{"title":"movie title","type":"string"}},"description":"Film (uuid,title)"},"Genre":{"title":"Genre","required":["uuid","name"],"type":"object","properties":{"uuid":{"title":"id","type":"string","format":"uuid"},"name":{"title":"genres name","type":"string"}},"description":"Movies genre"},"GenreManyResponse":{"title":"GenreManyResponse","required":["total","result"],"type":"object","properties":{"total":{"title":"Amount rows in source","type":"integer"},"result":{"title":"Result","type":"array","items":{"$ref":"#/components/schemas/Genre"}}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ImdbFilm":{"title":"ImdbFilm","required":["uuid","title","imdb_rating"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"title":{"title":"movie title","type":"string"},"imdb_rating":{"title":"IMDB rating","type":"number"}},"description":"Film (uuid,title, imdb)"},"ImdbfilmManyResponse":{"title":"ImdbfilmManyResponse","required":["total","result"],"type":"object","properties":{"total":{"title":"Amount rows in source","type":"integer"},"result":{"title":"Result","type":"array","items":{"$ref":"#/components/schemas/ImdbFilm"}}}},"Person":{"title":"Person","required":["uuid","full_name"],"type":"object","properties":{"uuid":{"title":"person id","type":"string","format":"uuid"},"full_name":{"title":"persons full name","type":"string"}},"description":"Person info"},"RoleMovies":{"title":"RoleMovies","required":["role","movies"],"type":"object","properties":{"role":{"title":"Role","type":"string"},"movies":{"title":"Movies","type":"array","items":{"$ref":"#/components/schemas/Film"}}}},"Sorting":{"title":"Sorting","enum":["+imdb_rating","-imdb_rating"],"description":"An enumeration."},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}